# cmd/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Noble blockchain indexer",
        "title": "Noble Indexer API",
        "termsOfService": "https://bakingbad.dev/terms",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/baking-bad/noble-indexer",
            "email": "hello@bakingbad.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/baking-bad/noble-indexer/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "noble.dipdup.net",
    "basePath": "/v1",
    "paths": {
...
        "/token": {
            "get": {
                "description": "Returns a paginated list of tokens (ERC20, ERC721, ERC1155). Can be filtered by token type or issuing contract address.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "List tokens",
                "operationId": "list-tokens",
                "parameters": [
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by contract address that issued the token",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of tokens to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of tokens to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ERC20",
                            "ERC721",
                            "ERC1155"
                        ],
                        "type": "string",
                        "description": "Filter by token standard (comma-separated list)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order by creation time (default: desc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tokens",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Token"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/token/{contract}/{token_id}": {
            "get": {
                "description": "Returns detailed information about a specific token including metadata, supply, and holder information. For ERC20 tokens use token_id=0, for ERC721/ERC1155 use the specific token ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token by contract and ID",
                "operationId": "get-token",
                "parameters": [
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Token contract address",
                        "name": "contract",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "0",
                        "description": "Token ID (use 0 for ERC20)",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token information",
                        "schema": {
                            "$ref": "#/definitions/responses.Token"
                        }
                    },
                    "204": {
                        "description": "Token not found"
                    },
                    "400": {
                        "description": "Invalid contract address or token ID",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/token_balance": {
            "get": {
                "description": "Returns a paginated list of token balances showing how many tokens each address holds. Can be filtered by address, contract, or token ID. Useful for finding token holders.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "List token balances",
                "operationId": "list-token-balances",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of balances to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of balances to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order (default: asc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by holder address",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by token contract address",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "0",
                        "description": "Filter by token ID",
                        "name": "token_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of token balances",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.TokenBalance"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/transfer": {
            "get": {
                "description": "Returns a paginated list of token transfer events (mint, burn, transfer). Can be filtered by transfer type, addresses, contract, token ID, transaction, block height, or time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "List token transfers",
                "operationId": "list-token-transfers",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of transfers to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of transfers to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order by timestamp (default: asc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 12345,
                        "description": "Filter by block height",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1692892095,
                        "description": "Filter by timestamp from (Unix timestamp)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1692892095,
                        "description": "Filter by timestamp to (Unix timestamp)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "burn",
                            "mint",
                            "transfer",
                            "unknown"
                        ],
                        "type": "string",
                        "description": "Filter by transfer type (comma-separated list)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "maxLength": 66,
                        "minLength": 66,
                        "type": "string",
                        "example": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef",
                        "description": "Filter by transaction hash (hexadecimal with 0x prefix)",
                        "name": "tx_hash",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by sender address",
                        "name": "address_from",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by recipient address",
                        "name": "address_to",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by token contract address",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "0",
                        "description": "Filter by token ID",
                        "name": "token_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of token transfers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Transfer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/transfer/{id}": {
            "get": {
                "description": "Returns detailed information about a specific token transfer event by its internal database ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token transfer by ID",
                "operationId": "get-token-transfer",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 12345,
                        "description": "Transfer internal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer information",
                        "schema": {
                            "$ref": "#/definitions/responses.Transfer"
                        }
                    },
                    "204": {
                        "description": "Transfer not found"
                    },
                    "400": {
                        "description": "Invalid transfer ID",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.Token": {
            "description": "Noble token information",
            "type": "object",
            "properties": {
                "contract": {
                    "type": "string",
                    "example": "0xdAC17F958D2ee523a2206206994597C13D831ec7"
                },
                "decimals": {
                    "type": "integer",
                    "example": 6
                },
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Tether USD"
                },
                "supply": {
                    "type": "string",
                    "example": "123456789"
                },
                "symbol": {
                    "type": "string",
                    "example": "USD"
                },
                "transfers_count": {
                    "type": "integer",
                    "example": 123
                },
                "type": {
                    "type": "string",
                    "example": "ERC20"
                }
            }
        },
        "responses.TokenBalance": {
            "description": "Token balance information",
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000001"
                },
                "contract": {
                    "type": "string",
                    "example": "0xdAC17F958D2ee523a2206206994597C13D831ec7"
                },
                "token_id": {
                    "type": "string",
                    "example": "0"
                },
                "value": {
                    "type": "string",
                    "example": "123456789"
                }
            }
        },
...
        "responses.Transfer": {
            "description": "Token transfer information",
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "123456789"
                },
                "contract": {
                    "type": "string",
                    "example": "0xdAC17F958D2ee523a2206206994597C13D831ec7"
                },
                "from": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000001"
                },
                "height": {
                    "type": "integer",
                    "example": 100
                },
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-01-01T01:01:01+00:00"
                },
                "to": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000002"
                },
                "token_id": {
                    "type": "string",
                    "example": "0"
                },
                "tx_hash": {
                    "type": "string",
                    "example": "0x0764012270afacd3b101bcfadaaa9fc3190d04ed90ff22c0ee59781e54858a7d"
                },
                "type": {
                    "type": "string",
                    "example": "transfer"
                }
            }
        }
    }
}